# src/main/java/ru/gb/timesheet/rest/ExceptionResponse.java
package ru.gb.timesheet.rest;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.Data;

@Data
@Schema(description = "Тип ответа с ошибкой")
public class ExceptionResponse {

  @Schema(description = "Текст ошибки")
  private String reason;

}